src/main/java/org/example/StringFormatter.java
package org.example;

public class StringFormatter {

    public static String capitalizeFirstLetter(String input) {
        if (input == null || input.isEmpty()) {
            return input;
        }
        return input.substring(0, 1).toUpperCase() + input.substring(1).toLowerCase();
    }

    public static String reverseString(String input) {
        if (input == null) {
            return null;
        }
        return new StringBuilder(input).reverse().toString();
    }

    public static String addExclamation(String input) {
        if (input == null) {
            return null;
        }
        return input + "!";
    }
}